pyproject.toml
[project]
name = "wf-engine"
version = "0.1.0"
description = "Default virtual environment for workflow engine"
requires-python = ">=3.12,<3.14"
dependencies = [
    "numpy>=2.0.0",
    "opencv-python>=4.10.0.84",
    "pillow>=10.4.0",
    "onnxruntime-gpu>=1.20.0",
    "openvino>=2024.4.0",
    "ultralytics>=8.3.0",
    "onnxruntime>=1.23.2",
]

[project.optional-dependencies]
agent = [
    "mcp>=1.1.0",
    "pyautogen>=0.2.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
